requirements.txt
connexion==2.2.0
cryptography==2.6.1
Flask==1.0.2
Flask-Cors==3.0.8
google-cloud-storage==1.16.1
google-cloud-datastore==1.8.0
gunicorn==19.9.0
jwkaas==0.3
MarkupSafe==1.1.1
python-dateutil==2.6.0
PyYAML==5.1.1
swagger-ui-bundle==0.0.2
setuptools >= 21.0.0
